requirements.txt
absl-py==0.15.0
alabaster==0.7.12
allennlp==0.9.0
antlr4-python3-runtime==4.8
attrs==21.2.0
Babel==2.9.1
biopython==1.79
blis==0.2.4
boto3==1.19.7
botocore==1.22.7
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
conllu==1.3.1
cycler==0.11.0
cymem==2.0.6
decord==0.6.0
docutils==0.17.1
editdistance==0.6.0
einops==0.3.2
fair-esm==0.4.0
flaky==3.7.0
Flask==2.0.2
Flask-Cors==3.0.10
ftfy==6.0.3
gevent==1.4.0
google-auth==2.3.2
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.41.1
h5py==2.10.0
idna==3.3
imagesize==1.2.0
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonnet==0.17.0
jsonpickle==2.0.0
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
murmurhash==1.0.6
nltk==3.6.5
numpy==1.21.3
numpydoc==1.1.0
oauthlib==3.1.1
omegaconf==2.1.1
overrides==3.1.0
packaging==21.2
pandas==1.3.4
parsimonious==0.8.1
Pillow==8.4.0
plac==0.9.6
pluggy==1.0.0
preshed==2.0.1
protobuf==3.19.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2021.3
PyYAML==6.0
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.15.0
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==1.0.1
scipy==1.7.1
sentencepiece==0.1.96
seqvec==0.4.1
six==1.16.0
snowballstemmer==2.1.0
spacy==2.1.9
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
srsly==1.0.5
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
thinc==7.0.8
threadpoolctl==3.0.0
toml==0.10.2
torch==1.10.0
tqdm==4.62.3
typing-extensions==3.10.0.2
typing-utils==0.1.0
Unidecode==1.3.2
urllib3==1.26.7
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==2.0.2
word2number==1.1
zipp==3.6.0
